dodano double

+++ main.c
@@ -4,7 +4,7 @@ int main(int args, char* argv[])
 {
     printf("Kamil psysa\n");
 
-    int i =7;
+    double i =7;
     //komentaz
 return 0;
 }
